svg-to-pdf/pyproject.toml
[tool.poetry]
name = "svg-to-pdf"
version = "0.1.0"
description = "Tool to convert SVG files to PDF with proper image support"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
svglib = "^1.5.1"
reportlab = "^4.1.0"
lxml = "^5.0.0"
cairosvg = "^2.7.1"
pillow = "^10.2.0"
weasyprint = "^61.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
svg-to-pdf = "svg_to_pdf.converter:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=svg_to_pdf --cov-report=xml --cov-report=term"

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true